Backend/app/requirements.txt
alembic
bcrypt
fastapi[standard]
jose
pyjwt
sqlmodel
passlib
pydantic_settings
pymysql
pytz